+++ NEWS.md
@@ -17,7 +17,8 @@
 - `reschola` is now armed with its own LimeSurvey API interface, providing several functions tailored to our needs:
     - `ls_call()` implements general and fully customisable call to the API (you must adhere to [the documentation](https://api.limesurvey.org/classes/remotecontrol_handle.html))
     - `ls_export()` is going to be used most of the time, as it fetches both participants and their responses, it basically merges outputs from  `ls_participants()`, and `ls_responses()`
-    - `ls_login()` is used for manual session key request; it may be never used, as every `ls_` call ensures the key is cached and not expired (key is stale after 2 hrs)
+    - `ls_login()` is used for manual session key request; **it may be never used directly**, as every `ls_` call ensures the key is cached and not expired (key is stale after 2 hrs)
+    - `ls_get_attrs()` for retrieval of the custom survey attributes and their "semantic" form (i.e. human-readable description)
     
     
 - new class `czech_date` with S3 method that prints the date in long format and with grammatically correct case, more in `as_czech_date()`
@@ -34,6 +35,7 @@
 
 - organization name change to correct form
 - `ggplot2::ScaleContinuousPosition` correctly imported, but still WIP
+- `ls_participants()` handles attributes correctly when used as a character vector
 
 
 # reschola 0.2.13